+++ lib/sleeping_king_studios/tools/string_tools.rb
@@ -58,6 +58,37 @@ module SleepingKingStudios::Tools
       plural_inflector.define_uncountable_word word
     end # method define_uncountable_word
 
+    # Adds the specified number of spaces to the start of each line of the
+    # string. Defaults to 2 spaces.
+    #
+    # @param str [String] The string to indent.
+    # @param count [Integer] The number of spaces to add.
+    #
+    # @return [String] The indented string.
+    def indent str, count = 2
+      str = require_string! str
+      pre = " " * count
+
+      map_lines(str) { |line| "#{pre}#{line}" }
+    end # method indent
+
+    # Yields each line of the string to the provided block and combines the
+    # results into a new multiline string.
+    #
+    # @param str [String] The string to map.
+    #
+    # @yieldparam line [String] The current line.
+    # @yieldparam index [Integer] The index of the current line.
+    #
+    # @return [String] The mapped string.
+    def map_lines str
+      str = require_string! str
+
+      str.each_line.with_index.reduce('') do |memo, (line, index)|
+        memo << yield(line, index)
+      end # each_line
+    end # method map_lines
+
     # Determines whether or not the given word is in plural form. If calling
     # #pluralize(word) is equal to word, the word is considered plural.
     #