+++ spec/sleeping_king_studios/tools/string_tools_spec.rb
@@ -193,6 +193,160 @@ RSpec.describe SleepingKingStudios::Tools::StringTools do
     end # it
   end # describe
 
+  describe '#indent' do
+    it { expect(instance).to respond_to(:indent).with(1..2).arguments }
+
+    describe 'with nil' do
+      it 'should raise an error' do
+        expect { described_class.indent nil }.to raise_error ArgumentError, /argument must be a string/
+      end # it
+    end # describe
+
+    describe 'with an empty string' do
+      it { expect(instance.indent '').to be == '' }
+    end # describe
+
+    describe 'with a single-line string' do
+      let(:string)   { 'Greetings, programs!' }
+      let(:expected) { "  #{string}" }
+
+      it { expect(instance.indent string).to be == expected }
+
+      describe 'with an indent count' do
+        let(:expected) { "    #{string}" }
+
+        it { expect(instance.indent string, 4).to be == expected }
+      end # describe
+    end # describe
+
+    describe 'with a single-line string with trailing newline' do
+      let(:string)   { "Greetings, programs!\n" }
+      let(:expected) { "  #{string}" }
+
+      it { expect(instance.indent string).to be == expected }
+
+      describe 'with an indent count' do
+        let(:expected) { "    #{string}" }
+
+        it { expect(instance.indent string, 4).to be == expected }
+      end # describe
+    end # describe
+
+    describe 'with a multi-line string' do
+      let(:string) do
+        "The Fellowship of the Ring\n"\
+        "The Two Towers\n"\
+        'The Return of the King'
+      end # let
+      let(:expected) do
+        "  The Fellowship of the Ring\n"\
+        "  The Two Towers\n"\
+        '  The Return of the King'
+      end # let
+
+      it { expect(instance.indent string).to be == expected }
+
+      describe 'with an indent count' do
+        let(:expected) do
+          "    The Fellowship of the Ring\n"\
+          "    The Two Towers\n"\
+          '    The Return of the King'
+        end # let
+
+        it { expect(instance.indent string, 4).to be == expected }
+      end # describe
+    end # describe
+  end # describe
+
+  describe '#map_lines' do
+    it { expect(instance).to respond_to(:map_lines).with(1).argument }
+
+    describe 'with nil' do
+      it 'should raise an error' do
+        expect { described_class.map_lines nil }.to raise_error ArgumentError, /argument must be a string/
+      end # it
+    end # describe
+
+    describe 'with an empty string' do
+      it { expect(instance.map_lines('') { |line| "- #{line}" }).to be == '' }
+    end # describe
+
+    describe 'with a single-line string' do
+      let(:string)   { 'Greetings, programs!' }
+      let(:expected) { "- #{string}" }
+
+      it 'should map each line' do
+        expect(instance.map_lines(string) { |line| "- #{line}" }).
+          to be == expected
+      end # it
+
+      describe 'with a block that yields the index' do
+        let(:expected) { "0. #{string}" }
+
+        it 'should map each line' do
+          expect(instance.map_lines(string) do |line, index|
+            "#{index}. #{line}"
+          end). # map_lines
+            to be == expected
+        end # it
+      end # describe
+    end # describe
+
+    describe 'with a single-line string with trailing newline' do
+      let(:string)   { "Greetings, programs!\n" }
+      let(:expected) { "- #{string}" }
+
+      it 'should map each line' do
+        expect(instance.map_lines(string) { |line| "- #{line}" }).
+          to be == expected
+      end # it
+
+      describe 'with a block that yields the index' do
+        let(:expected) { "0. #{string}" }
+
+        it 'should map each line' do
+          expect(instance.map_lines(string) do |line, index|
+            "#{index}. #{line}"
+          end). # map_lines
+            to be == expected
+        end # it
+      end # describe
+    end # describe
+
+    describe 'with a multi-line string' do
+      let(:string) do
+        "The Fellowship of the Ring\n"\
+        "The Two Towers\n"\
+        'The Return of the King'
+      end # let
+      let(:expected) do
+        "- The Fellowship of the Ring\n"\
+        "- The Two Towers\n"\
+        '- The Return of the King'
+      end # let
+
+      it 'should map each line' do
+        expect(instance.map_lines(string) { |line| "- #{line}" }).
+          to be == expected
+      end # it
+
+      describe 'with a block that yields the index' do
+        let(:expected) do
+          "0. The Fellowship of the Ring\n"\
+          "1. The Two Towers\n"\
+          '2. The Return of the King'
+        end # let
+
+        it 'should map each line' do
+          expect(instance.map_lines(string) do |line, index|
+            "#{index}. #{line}"
+          end). # map_lines
+            to be == expected
+        end # it
+      end # describe
+    end # describe
+  end # describe
+
   describe '#plural?' do
     it { expect(instance).to respond_to(:plural?).with(1).argument }